Communication redesign...
The communication packet is getting too bulky wasting context window.
Redisigning communication protocol so that only payloads that are used
are sent on the wire. These are wrapped in a lightweight communication
packet that can now hold any amount of payload combinations.

The router will now be split in two, one for handlers, one for AI models.
AI Models and handlers will now both be Destinations and use the Strategy
design pattern.